Move session recovery popup to home screen
