+++ main.kv
@@ -42,7 +42,7 @@ RootUI:
     PreviousWorkoutsScreen:
         name: "previous_workouts"
 
-<HomeScreen@MDScreen>:
+<HomeScreen>:
     md_bg_color: PINK_BG
     ScrollView:
         do_scroll_x: False
