+++ tests/test_ui.py
@@ -63,8 +63,8 @@ def test_switch_tab_updates_current_tab():
 
 
 @pytest.mark.skipif(not kivy_available, reason="Kivy and KivyMD are required")
-def test_welcome_screen_triggers_recovery(monkeypatch):
-    from ui.screens.general.welcome_screen import WelcomeScreen
+def test_home_screen_triggers_recovery(monkeypatch):
+    from ui.screens.general.home_screen import HomeScreen
     from backend.workout_session import WorkoutSession
 
     dummy_session = object()
@@ -79,12 +79,36 @@ def test_welcome_screen_triggers_recovery(monkeypatch):
     def fake_dialog(self, session):
         called["session"] = session
 
-    monkeypatch.setattr(WelcomeScreen, "_show_recovery_dialog", fake_dialog)
+    monkeypatch.setattr(HomeScreen, "_show_recovery_dialog", fake_dialog)
 
-    screen = WelcomeScreen()
+    screen = HomeScreen()
     screen.on_enter()
 
     assert called.get("session") is dummy_session
+
+
+@pytest.mark.skipif(not kivy_available, reason="Kivy and KivyMD are required")
+def test_home_screen_no_recovery_shows_info(monkeypatch):
+    from ui.screens.general.home_screen import HomeScreen
+    from backend.workout_session import WorkoutSession
+
+    monkeypatch.setattr(
+        WorkoutSession,
+        "load_from_recovery",
+        classmethod(lambda cls: None),
+    )
+
+    called = {}
+
+    def fake_dialog(self):
+        called["called"] = True
+
+    monkeypatch.setattr(HomeScreen, "_show_no_session_dialog", fake_dialog)
+
+    screen = HomeScreen()
+    screen.on_enter()
+
+    assert called.get("called") is True
 
 
 @pytest.mark.skipif(not kivy_available, reason="Kivy and KivyMD are required")
